# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/send": {
            "post": {
                "description": "Этот эндпоит отправляет сообщение на сервер",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "отправка сообщения",
                "parameters": [
                    {
                        "description": "Сообщение для отправки",
                        "name": "message",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Message"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешная отправка",
                        "schema": {
                            "$ref": "#/definitions/model.MessageResponse"
                        }
                    },
                    "400": {
                        "description": "Ошибка запроса",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/transfer": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "передача сегментов в кафку",
                "parameters": [
                    {
                        "description": "Сегмент",
                        "name": "message",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Segment"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешно получено и отправлено в кафку",
                        "schema": {
                            "$ref": "#/definitions/model.MessageResponse"
                        }
                    },
                    "400": {
                        "description": "Ошибка запроса",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "model.Message": {
            "type": "object",
            "properties": {
                "sender_name": {
                    "type": "string"
                },
                "string_message": {
                    "type": "string"
                }
            }
        },
        "model.MessageResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "model.Segment": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "payload": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "segment_number": {
                    "type": "integer"
                },
                "sender_name": {
                    "type": "string"
                },
                "total_segments": {
                    "type": "integer"
                }
            }
        }
    }
}